backend/requirements.txt
fastapi
uvicorn
scikit-learn
numpy
joblib
pandas
prophet
torch
torchvision
timm
Pillow
python-multipart
groq
langchain-chroma
langchain-huggingface
chromadb